+++ NEWS.md
@@ -49,6 +49,7 @@
 * Updating roxygen dependencies to correctly document the package.
 * Extend documentation and testing of `set_diff`
 * The default value for `conf_int` in `qacf` has been modified from `0.05` to `1 - getOption("qwraps2_alpha", 0.05)`
+* `ggplot2_extract_legend` is now and S3 method
 
 # qwraps2 0.5.2